In which our hero saved some memory with the correct data type

+++ src/helper/linkedlist.h
@@ -24,7 +24,7 @@ typedef struct {
     void *prev;
     void *next;
     char *name;
-    int active;
+    uint8_t active;
 } linkedlist_t;
 
 void *linkedlist_appendElement(void *list, void *element);
